refactor: 공통 스타일 ItemsSection.module.css을 best, current기준으로 분리
