+++ src/pages/ItemsPage/sections/CurrentItemsSection.module.css
@@ -1,0 +1,97 @@
+.cards-section {
+  display: flex;
+  flex-direction: column;
+  max-width: 1250px;
+  width: 100%;
+  margin: 0 auto;
+  padding: 0 24px;
+}
+
+.cards-section.best {
+  gap: 16px;
+}
+
+.cards-section.current {
+  gap: 24px;
+}
+
+.section-header-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  gap: 12px;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.section-title {
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 32px;
+  color: var(--color-gray900);
+  flex-grow: 1;
+  order: var(--search-title-order);
+  min-width: 200px;
+}
+
+.search-input-container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  background-color: var(--color-gray200);
+  width: var(--search-input-container-width);
+  padding: 0 9px;
+  height: 42px;
+  border-radius: 12px;
+  order: var(--search-input-order);
+  flex-grow: var(--search-input-flex-grow);
+}
+
+.search-input-icon {
+  aspect-ratio: 1/1;
+}
+
+.search-input {
+  border: none;
+  background-color: var(--color-gray200);
+  flex-grow: 1;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 26px;
+  outline: none;
+}
+
+.search-input::placeholder {
+  color: var(--color-gray400);
+}
+
+.search-submit {
+  border-radius: 8px;
+  font-weight: 600;
+  font-size: 16px;
+  line-height: 26px;
+  padding: 0 23px;
+  color: var(--color-gray100);
+  height: 42px;
+  order: var(--search-submit-order);
+}
+
+.search-select {
+  width: var(--search-select-width);
+  padding: 0 20px;
+  border-radius: 12px;
+  border: 1px solid var(--color-gray300);
+  appearance: none; /* 기본 브라우저 UI 제거 */
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  background-image: var(--search-select-image);
+  background-repeat: no-repeat;
+  background-position: var(--search-select-image-position);
+  background-size: 24px;
+  color: var(--color-gray800);
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 26px;
+  height: 42px;
+  order: var(--search-select-order);
+}
